Split stimulus controller into concerns
I think this might make more sense than outlets or helper classes.

An outlet might make sense for the dialog behavior.
But all this behavior does feel part of the same whole at this point.

Pulling out helper classes would work, too. But it seems like the state we're carrying is fine to live inside a Combobox class (well, technically an object in this implementation).

I've never architected JS like this. Seems like a good opportunity to see if this form of modeling carries weight.