Add enum support
Closes https://github.com/josefarias/hotwire_combobox/issues/32

In addition to passing an `AR::Relation`, an array of hashes, or a url/path string;
we now also support passing a hash, an array of strings, or an array of arrays.
This is specially useful for supporting enums.

Here's how the combobox will behave with each of the new supported types:

* Hash: hash is iterated with `#map`. For each item, options are determined as `{ display: option.first, value: option.last }`
* Array of strings: For each item, options are determined as `{ display: option, value: option }`
* Array of arrays: For each item, options are determined as `{ display: option.first, value: option.last }`