Fix filtering inconsistencies
Closes https://github.com/josefarias/hotwire_combobox/pull/27

This PR fixes two things:

1. If you filter the listbox by typing an option all the way through,
the list is still filtered after closing and reopening it.
This was not the case when clicking an option without typing.
Now, selecting an option also filters the list for a consistent experience.

2. _isValidNewOption optionally ignores autocomplete sufficiency to allow
for new options which start the same as existing ones. But exact matches
shouldn't ever be ignored. Exact matches are never new.